Modify dataTransfer and imageTest to be compatible with large memory operations
On TITAN, the END_SIZE is set to 2 GB. This is ok for an unsigned int
to store this value. But when in a loop, the looper will be 0 at the
end of the loop if the looper is not set to unsigned long long. Thus, we
just choose to use unsigned long long for everthing that has something
to do with END_SIZE.